chore(news-list): move to components
