+++ src/components/news-list/styled.ts
